Split the SpotifyAccount god-class into focused mixins
account.py had grown to 1319 lines and 52 methods on a single class,
tripping pylint C0302 (too-many-lines) and R0904 (too-many-public-
methods). It was also the file every new feature landed in, making it
the largest merge-conflict surface in the codebase.

Convert the module into a package whose __init__ stays the public
import path and acts as a facade: it keeps the class constants, the
constructor, base_refresh_rate, async_from_config_entry, and the shared
instance state. The behaviour is assembled from six single-
responsibility mixins that operate purely on that state:

  - PagingMixin      dataset access and Spotify pagination
  - TokenMixin       token and session lifecycle
  - ProfileMixin     identity and profile
  - LibraryMixin     content and library reads
  - PlaybackMixin    device and playback control
  - InternalApiMixin unofficial internal-endpoint reads

Pure move of the method bodies, no logic changes; the public API is
identical so the 32 consumer files are untouched. Only three test
patch targets move with the symbols they mock (run_coroutine_threadsafe
-> token, async_get_playlist_length/name -> internal).

Every resulting file is grade A on the Maintainability Index and under
350 lines (was 1319). C0302 and R0904 are cleared. R0902 remains on the
facade by design, since it is the single owner of account state, and is
annotated as such.
