--- custom_components/spotcast/spotify/account.py
@@ -1,1319 +1,0 @@
-"""Module for the spotify account class.
-
-Classes:
-    SpotifyAccount
-"""
-
-from asyncio import (
-    run_coroutine_threadsafe,
-    Lock,
-    sleep,
-    TimeoutError,
-)
-from copy import deepcopy
-import datetime as dt
-from logging import getLogger
-from time import time
-from typing import Any
-
-from spotipy import Spotify, SpotifyException
-from homeassistant.core import HomeAssistant
-from homeassistant.config_entries import ConfigEntry, SOURCE_REAUTH
-from homeassistant.helpers.device_registry import DeviceInfo, DeviceEntryType
-from homeassistant.helpers.storage import Store
-
-from custom_components.spotcast.const import DOMAIN
-from custom_components.spotcast.sessions import (
-    DesktopSession,
-    PublicSession,
-    ConnectionSession,
-    async_get_config_entry_implementation,
-)
-from custom_components.spotcast.utils import ensure_default_data, copy_to_dict
-from custom_components.spotcast.spotify.dataset import Dataset
-from custom_components.spotcast.spotify.internal_api import (
-    async_get_playlist_length,
-    async_get_playlist_name,
-)
-from custom_components.spotcast.spotify.search_query import SearchQuery
-from custom_components.spotcast.spotify.utils import select_image_url
-from custom_components.spotcast.spotify.exceptions import (
-    PlaybackError,
-    TokenError,
-)
-from custom_components.spotcast.entry_data import ApiItem
-from custom_components.spotcast.sessions.exceptions import (
-    InternalServerError,
-    UpstreamServerNotready,
-)
-
-
-LOGGER = getLogger(__name__)
-
-
-class SpotifyAccount:
-    """The account of a Spotify user.
-
-    Attributes:
-        - hass(HomeAssistant): The Home Assistance instance
-        - sessions(dict[str, ConnectionSession]): A dictionary with
-            both the Internal (Private) and External (Public) API
-            access
-        - is_defaults(bool): set to True if the account should be
-            treated as default when calling services
-
-    Properties:
-        - id(str): the identifier of the account
-        - name(str): the display name for the account
-        - profile(dict): the full profile dictionary of the account
-        - country(str): the country code where the account currently
-            is.
-        - image_link(str): the profile image image link
-        - product(str): the current subscription product the user has
-        - type(str): the type of account loaded
-        - liked_songs_uri(str): the uri for the liked_songs playlist
-
-    Constants:
-        - SCOPE(tuple): A list of API permissions required for the
-            instance to work properly
-        - DJ_URI(str): the uri for the DJ playlist
-        - REFRESH_RATE(int): default rate at which to deem the cache
-            deprecated
-       - DATASETS(dict[str, dict]): Default configuration for datasets
-            used by the account
-
-    Methods:
-        - get_profile_value
-        - get_token
-        - async_get_token
-        - async_connect
-        - async_ensure_tokens_valid
-        - async_profile
-        - async_devices
-        - async_playlists
-        - async_wait_for_devices
-        - async_register_chromecast_player
-        - async_play_media
-        - async_apply_extras
-        - async_shuffle
-        - async_liked_songs
-        - async_like_media
-        - async_repeat
-        - async_set_volume
-
-    Functions:
-        - async_from_config_entry
-    """
-
-    SCOPE = (
-        "user-modify-playback-state",
-        "user-read-playback-state",
-        "user-read-private",
-        "playlist-read-private",
-        "playlist-read-collaborative",
-        "user-library-read",
-        "user-library-modify",
-        "user-top-read",
-        "user-read-playback-position",
-        "user-read-recently-played",
-        "user-follow-read",
-    )
-
-    DJ_URI = "spotify:playlist:37i9dQZF1EYkqdzj48dyYq"
-    REFRESH_RATE = 30
-    DATASETS = {
-        "devices": {
-            "refresh_factor": 1,
-            "can_expire": False,
-        },
-        "liked_songs": {
-            "refresh_factor": 10,
-            "can_expire": False,
-        },
-        "playlists": {
-            "refresh_factor": 10,
-            "can_expire": False,
-        },
-        "profile": {
-            "refresh_factor": 20,
-            "can_expire": True,
-        },
-        "categories": {
-            "refresh_factor": 20,
-            "can_expire": False,
-        },
-        "playback_state": {
-            "refresh_factor": 1 / 2,
-            "can_expire": False,
-        },
-    }
-
-    SESSION_CONFIG_MAP = {
-        "public": "external_api",
-        "private": "desktop_api",
-    }
-
-    def __init__(
-        self,
-        entry_id: str,
-        hass: HomeAssistant,
-        public_session: PublicSession,
-        private_session: DesktopSession,
-        is_default: bool = False,
-        base_refresh_rate: int = 30,
-    ):
-        """The account of a Spotify user.
-
-        Args:
-            entry_id(str): The id of the config entry in HomeAssistant
-                for the account.
-            hass(HomeAssistant): The Home Assistant Instance
-            public_session(OAuth2Session): The public api session
-                for the Spotify Account
-            private_session(InternalSession): The private api
-                session for the Spotify Account
-            is_default(bool, optional): True if account is treated as
-                default for service call. Defaults to False.
-            base_refresh_rate(int, optional): The base refresh rate
-                used to update dateset
-        """
-        self.entry_id = entry_id
-        self.hass = hass
-        self._lock = Lock()
-        self.sessions: dict[str, ConnectionSession] = {
-            "public": public_session,
-            "private": private_session,
-        }
-        self.apis: dict[str, Spotify] = {}
-
-        for name, session in self.sessions.items():
-            self.apis[name] = Spotify(auth=session.access_token)
-
-        self.is_default = is_default
-        self._base_refresh_rate = base_refresh_rate
-
-        self._datasets: dict[str, Dataset] = {}
-        self._last_playback_state = {}
-        self._context_name_cache: tuple[str | None, str | None] = (None, None)
-        self._playback_store = Store(
-            hass,
-            1,
-            f"spotcast_{entry_id}_last_state",
-        )
-
-        for name, dataset in self.DATASETS.items():
-            refresh_rate = dataset["refresh_factor"] * self._base_refresh_rate
-            can_expire = dataset["can_expire"]
-            self._datasets[name] = Dataset(name, refresh_rate, can_expire)
-
-    @property
-    def base_refresh_rate(self) -> int:
-        """Returns the current base refresh rate."""
-        return self._base_refresh_rate
-
-    @base_refresh_rate.setter
-    def base_refresh_rate(self, value: int):
-        """Sets the base refresh rate and updates the dataset.
-
-        Args:
-            value(int): the new base refresh_rate
-        """
-        self._base_refresh_rate = value
-
-        for name, dataset in self._datasets.items():
-            refresh_factor = self.DATASETS[name]["refresh_factor"]
-            dataset.refresh_rate = value * refresh_factor
-
-    @property
-    def id(self) -> str:
-        """Returns the id of the account."""
-        return self.get_profile_value("id")
-
-    @property
-    def name(self) -> str:
-        """Returns the name of the account."""
-        name = self.get_profile_value("display_name")
-
-        if name is None:
-            name = self.id
-
-        return name
-
-    @property
-    def profile(self) -> dict:
-        """Returns the full profile dictionary of the account."""
-        return self.get_dataset("profile")
-
-    @property
-    def playlists(self) -> list:
-        """Returns the list of playlists for the account."""
-        return self.get_dataset("playlists")
-
-    @property
-    def categories(self) -> list:
-        """Returns the list of Browse categories for the account."""
-        return self.get_dataset("categories")
-
-    @property
-    def liked_songs(self) -> list:
-        """Returns the list of liked songs for the account."""
-        liked_songs = self.get_dataset("liked_songs")
-        liked_songs = [x["track"]["uri"] for x in liked_songs]
-        return liked_songs
-
-    @property
-    def devices(self) -> list:
-        """Returns the list of devices linked to the account."""
-        return self.get_dataset("devices")
-
-    @property
-    def playback_state(self) -> list:
-        """Returns the list of devices linked to the account."""
-        return self.get_dataset("playback_state")
-
-    @property
-    def country(self) -> str:
-        """Returns the current country in which the account resides."""
-        return self.get_profile_value("country")
-
-    @property
-    def image_link(self) -> str:
-        """Returns the link for the account profile image."""
-        images = self.get_profile_value("images")
-        return select_image_url(images)
-
-    @property
-    def product(self) -> str:
-        """Returns the account subscription product."""
-        return self.get_profile_value("product")
-
-    @property
-    def type(self) -> str:
-        """Returns the type of account."""
-        return self.get_profile_value("type")
-
-    @property
-    def liked_songs_uri(self) -> str:
-        """Returns the liked songs uri for the account."""
-        return f"spotify:user:{self.id}:collection"
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Returns the Home Assistant device info of the account."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self.id)},
-            manufacturer="Spotify AB",
-            model=f"Spotify {self.profile['product']}",
-            name=f"Spotcast {self.name}",
-            entry_type=DeviceEntryType.SERVICE,
-            configuration_url="https://open.spotify.com",
-        )
-
-    @property
-    def health_status(self) -> dict[str, bool]:
-        """Returns the health status of the underlying sessions."""
-        health = {}
-
-        for key, session in self.sessions.items():
-            health[key] = session.is_healthy
-
-        return health
-
-    @property
-    def active_device(self) -> str:
-        """Returns the current active device."""
-        playback_state = self.playback_state
-
-        if self.playback_state == {}:
-            return None
-
-        return playback_state["device"]["id"]
-
-    def get_profile_value(self, attribute: str) -> Any:
-        """Returns the value for a profile element.
-
-        Args:
-            attribute(str): the attribute to fetch from the profile
-
-        Raises:
-            ProfileNotLoadedError: Raised if the profile hasn't been
-                loaded yet.
-
-        Returns:
-            Any: the value at the key in the profile
-        """
-        profile = self._datasets["profile"].data
-
-        return profile.get(attribute)
-
-    def get_dataset(self, name: str) -> list | dict:
-        """Retrieves a specific dataset."""
-        return self._datasets[name].data
-
-    def get_token(self, api: str) -> str:
-        """Retrives a token from the requested session.
-
-        Args:
-            api(str): The api to retrieve from. Cann be `internal`
-                or `external`.
-
-        Returns:
-            str: token for the requested session
-        """
-        return run_coroutine_threadsafe(
-            self.async_get_token(api), self.hass.loop
-        ).result()
-
-    async def async_get_token(self, api: str) -> str:
-        """Retrives a token from the requested session.
-
-        Args:
-            api(str): The api to retrieve from. Can be `internal` or
-                `external`.
-
-        Returns:
-            - str: token for the requested session
-        """
-        await self.sessions[api].async_ensure_token_valid()
-        return self.sessions[api].access_token
-
-    async def async_get_internal_playlist_length(
-        self, uri: str
-    ) -> int | None:
-        """Returns the track count of a playlist through Spotify's
-        unofficial internal endpoint, which also resolves editorial and
-        algorithmic playlists that 404 on the public Web API.
-
-        Args:
-            - uri(str): the playlist uri
-
-        Returns:
-            - int | None: the track count, or None when the internal
-                endpoint is unavailable (callers should fall back)
-        """
-        return await async_get_playlist_length(self, uri)
-
-    async def async_get_current_context_name(self) -> str | None:
-        """Returns the human-readable name of the playlist currently
-        playing on the account.
-
-        The context uri comes from the public playback state, but its
-        name is resolved through the unofficial internal endpoint so it
-        also works for editorial/algorithmic playlists. The last
-        resolved name is cached per context uri to avoid a lookup on
-        every refresh.
-
-        Returns:
-            - str | None: the playlist name, or None when nothing is
-                playing, the context is not a playlist, or the name
-                could not be resolved
-        """
-        playback = self.playback_state
-        context = playback.get("context") if isinstance(playback, dict) else None
-        uri = context.get("uri") if isinstance(context, dict) else None
-
-        if not uri or ":playlist:" not in uri:
-            return None
-
-        if uri == self._context_name_cache[0]:
-            return self._context_name_cache[1]
-
-        name = await async_get_playlist_name(self, uri)
-
-        if name is not None:
-            self._context_name_cache = (uri, name)
-
-        return name
-
-    async def async_ensure_tokens_valid(
-        self,
-        skip_profile: bool = False,
-        reauth_on_fail: bool = True,
-        force_entry_update: bool = False,
-    ):
-        """Ensures the token are valid.
-
-        Args:
-            skip_profile(bool, optional): set True to skip the
-                profile update. Defaults to False
-            reauth_on_fail(bool, optional): Asks for reauthorisation
-                of the entry on failure to get token. Defaults to True.
-            force_entry_update(bool, optional): Forces the config entry
-                update even if no data was updated.
-        """
-        if not skip_profile:
-            await self.async_profile()
-
-        async with self._lock:
-            entry = self.hass.config_entries.async_get_entry(self.entry_id)
-            new_data = deepcopy({**entry.data})
-            need_update = False
-
-            for key, session in self.sessions.items():
-                was_updated = await self._async_refresh_single_token(
-                    session,
-                    reauth_on_fail,
-                )
-
-                LOGGER.debug(
-                    "Session `%s` token updated: %s",
-                    key,
-                    was_updated,
-                )
-
-                need_update = need_update or was_updated
-
-                new_data[self.SESSION_CONFIG_MAP[key]] = {**session.data}
-
-                self.apis[key].set_auth(session.access_token)
-
-            if need_update or force_entry_update:
-                LOGGER.info(
-                    "New information receive, updating entry for `%s`",
-                    self.entry_id,
-                )
-                new_data["last_update"] = dt.datetime.now().isoformat()
-
-                result = self.hass.config_entries.async_update_entry(
-                    entry=entry,
-                    data=new_data,
-                )
-
-                LOGGER.debug("Config Entry updated: %s", result)
-
-    async def _async_refresh_single_token(
-        self,
-        session: ConnectionSession,
-        reauth_on_fail: bool = True,
-    ) -> bool:
-        """Refreshes the token a a specific session."""
-        try:
-            was_updated = await session.async_ensure_token_valid()
-            return was_updated
-        except TokenError as exc:
-            if reauth_on_fail:
-                LOGGER.error(
-                    "An error occured while trying to refresh the "
-                    "token. Reauthentication requested"
-                )
-
-                entry = self.hass.config_entries.async_get_entry(self.entry_id)
-                entry.async_start_reauth(
-                    self.hass, context={"source": SOURCE_REAUTH}
-                )
-
-            raise exc
-
-    def _request_reauth(self, entry: ConfigEntry):
-        """Requests reauth for an entry."""
-        entry.async_start_reauth(
-            self.hass,
-            context={"source": SOURCE_REAUTH},
-        )
-
-    async def async_profile(self, force: bool = False) -> dict:
-        """Test the connection and returns a user profile.
-
-        Args:
-            force(bool, optional): Forces the profile update if True.
-                Defaults to False
-
-        Returns:
-            dict: the raw profile dictionary from the Spotify API
-        """
-        await self.async_ensure_tokens_valid(skip_profile=True)
-        LOGGER.debug("Getting Profile from Spotify")
-
-        dataset = self._datasets["profile"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing profile dataset")
-                data = await self.hass.async_add_executor_job(
-                    self.apis["public"].me
-                )
-                dataset.update(data)
-            else:
-                LOGGER.debug("Using cached profile dataset")
-
-        return self.profile
-
-    async def async_devices(self, force: bool = False) -> list[dict]:
-        """Returns the list of devices."""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Getting Devices for account `%s`", self.name)
-
-        dataset = self._datasets["devices"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing devices dataset")
-                data = await self.hass.async_add_executor_job(
-                    self.apis["public"].devices
-                )
-                dataset.update(data["devices"])
-            else:
-                LOGGER.debug("Using Cached devices dataset")
-
-        return self.devices
-
-    async def async_saved_episodes(
-        self,
-        limit: int = None,
-    ) -> list[dict]:
-        """Retrieves the list of podcast episode saved.
-
-        Args:
-            limit(int, optional): If not None, stops getting episode
-                once the limit of item reached. Defaults to None
-
-        Return:
-            - list[dict]: a list of dictionary with the episodes
-                information
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug(
-            "Getting List of saved podcast episode for account `%s`",
-            self.name,
-        )
-
-        return await self._async_pager(
-            self.apis["public"].current_user_saved_episodes,
-            appends=[self.country],
-            max_items=limit,
-        )
-
-    async def async_get_track(self, uri: str) -> dict:
-        """Retrieves track information.
-
-        Args:
-            uri(str): The URI of the track to search
-
-        Returns:
-            dict: the songs details
-        """
-        await self.async_ensure_tokens_valid()
-
-        LOGGER.debug("Getting track information for `%s`", uri)
-
-        result = await self.hass.async_add_executor_job(
-            self.apis["public"].track, uri, self.country
-        )
-
-        return result
-
-    async def async_get_playlist(self, uri: str) -> dict:
-        """Retrieves a playlist information.
-
-        Args:
-            uri(str): the URI of the playlist to search
-
-        Returns:
-            dict: the playlist details
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching information from playlist `%s`", uri)
-
-        playlist_id = self._id_from_uri(uri)
-
-        result = await self.hass.async_add_executor_job(
-            self.apis["public"].playlist,
-            playlist_id,
-            None,
-            self.country,
-        )
-
-        return result
-
-    async def async_get_album(self, uri: str) -> dict:
-        """Retrieves an album information.
-
-        Args:
-            uri(str): the URI of the album to search
-
-        Returns:
-            dict: the album details
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching information for album `%s`", uri)
-
-        album_id = self._id_from_uri(uri)
-
-        result = await self.hass.async_add_executor_job(
-            self.apis["public"].album,
-            album_id,
-            self.country,
-        )
-
-        return result
-
-    async def async_get_artist_top_tracks(self, uri: str) -> list[dict]:
-        """Retrieves the list of top tracks for an artist.
-
-        Args:
-            uri(str): the URI of the artist to search
-
-        Returns:
-            list[dict]: the list of top songes for an artist
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching Top Tracks for artist `%s`", uri)
-
-        result = await self.hass.async_add_executor_job(
-            self.apis["public"].artist_top_tracks,
-            uri,
-            self.country,
-        )
-
-        return result["tracks"]
-
-    async def async_get_playlist_tracks(self, uri: str) -> list[dict]:
-        """Retrieves the list of tracks inside a playlist."""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching tracks from playlist `%s`", uri)
-
-        playlist_id = self._id_from_uri(uri)
-
-        result = await self._async_pager(
-            function=self.apis["public"].playlist_tracks,
-            prepends=[playlist_id, None],
-            appends=[self.country],
-        )
-
-        return result
-
-    @staticmethod
-    def _id_from_uri(uri: str) -> str:
-        """extracts the id from a uri"""
-        return uri.rsplit(":", maxsplit=1)[-1]
-
-    async def async_get_show_episodes(
-        self, uri: str, limit: int = None
-    ) -> list[dict]:
-        """Retrieves the list of episodes for a podcast show.
-
-        Args:
-            - uri(str): the uri of the spotify podcast show to call
-            - limit(int, optional): limit the number of items to
-                retrieve. Retrives all episodes if None. Defaults to
-                None.
-
-        Returns:
-            - list[dict]: list of dictionaries with podcast episodes
-                information
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching episodes from show `%s`", uri)
-
-        result = await self._async_pager(
-            function=self.apis["public"].show_episodes,
-            prepends=[uri],
-            appends=[self.country],
-            max_items=limit,
-        )
-
-        return result
-
-    async def async_get_episode(self, uri: str) -> str:
-        """Retrieves the information of a podcast episode"""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Fetching information from episode `%s`")
-
-        result = await self.hass.async_add_executor_job(
-            self.apis["public"].episode,
-            uri,
-            self.country,
-        )
-
-        return result
-
-    async def async_last_playback_state(self) -> dict:
-        """The last known playback state"""
-
-        if self._last_playback_state != {}:
-            return self._last_playback_state
-
-        return await self._playback_store.async_load() or {}
-
-    async def async_playback_state(self, force: bool = False) -> dict:
-        """Returns the current playback state"""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Getting Playback Sate for account `%s`", self.name)
-
-        dataset = self._datasets["playback_state"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing playback state dataset")
-                data = await self.hass.async_add_executor_job(
-                    self.apis["public"].current_playback,
-                    self.country,
-                    "episode",
-                )
-
-                if data is None:
-                    data = {}
-                else:
-                    self._last_playback_state = data
-                    await self._playback_store.async_save(data)
-
-                dataset.update(data)
-            else:
-                LOGGER.debug("Using Cached playback state dataset")
-
-        return self.playback_state
-
-    async def async_playlists_count(self) -> int:
-        """Returns the number of user playlist for an account."""
-        await self.async_ensure_tokens_valid()
-
-        return await self._async_get_count(
-            self.apis["public"].current_user_playlists
-        )
-
-    async def async_playlists(
-        self,
-        force: bool = False,
-        max_items: int = None,
-    ) -> list[dict]:
-        """Returns a list of playlist for the current user."""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Getting Playlist for account `%s`", self.name)
-
-        dataset = self._datasets["playlists"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing playlists dataset")
-
-                playlists = await self._async_pager(
-                    self.apis["public"].current_user_playlists,
-                    max_items=max_items,
-                )
-
-                dataset.update(playlists)
-
-            else:
-                LOGGER.debug("Using cached playlists dataset")
-
-        return self.playlists
-
-    async def async_search(
-        self,
-        query: SearchQuery,
-        limit: int = 20,
-    ) -> list[dict]:
-        """Makes a search query and returns the result.
-
-        Args:
-            query(SearchQuery): The search query to run
-            limit(int, optional): The maximum amount of item to
-                retrieve in each category. Defaults to 20.
-        """
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug(
-            "Getting Search Result `%s` for account `%s`",
-            query.search,
-            self.name,
-        )
-
-        search_result = await self.hass.async_add_executor_job(
-            self.apis["public"].search,
-            query.query_string,
-            limit,
-            0,
-            query.item_types_string,
-            self.country,
-        )
-
-        result = {}
-
-        for item_type in query.item_types:
-            key = f"{item_type}s"
-            result[key] = [
-                item
-                for item in (search_result[key]["items"] or [])
-                if item is not None
-            ]
-
-        return result
-
-    async def async_wait_for_device(self, device_id: str, timeout: int = 12):
-        """Asycnhronously wait for a device to become available.
-
-        Args:
-            device_id(str): the spotify id of the device to wait for
-            timeout(int): the timeout delay to wait for before
-                raising an error.
-
-        Raises:
-            - TimeoutError: raised when waiting for the device goes
-                beyond the set delay
-        """
-        LOGGER.debug("Waiting for device `%s` to become available", device_id)
-
-        end_time = time() + timeout
-
-        while time() <= end_time:
-            devices = await self.async_devices(force=True)
-            devices = {x["id"]: x for x in devices}
-
-            try:
-                devices[device_id]
-                return
-            except KeyError:
-                LOGGER.debug("Device `%s` not yet available", device_id)
-                await sleep(timeout / 4)
-
-        raise TimeoutError(
-            f"device `{device_id}` still not available after {timeout} sec."
-        )
-
-    async def async_apply_extras(
-        self,
-        device_id: str,
-        extras: dict,
-    ):
-        """Applies extra settings on an account.
-
-        Args:
-            account(SpotifyAccount): the account to apply extras to
-            device_id(str): the device to set the extras to
-            extras(dict): the extra settings to apply
-        """
-        actions = {
-            "volume": self.async_set_volume,
-            "shuffle": self.async_shuffle,
-            "repeat": self.async_repeat,
-        }
-
-        for key, value in extras.items():
-            if key not in actions:
-                continue
-
-            try:
-                await actions[key](value, device_id)
-            except SpotifyException as exc:
-                LOGGER.warning(
-                    "Could not apply %s=%s to device %s: %s",
-                    key,
-                    value,
-                    device_id,
-                    exc.msg,
-                )
-
-    async def async_play_media(
-        self,
-        device_id: str,
-        context_uri: str = None,
-        uris: list[str] = None,
-        offset: int | str = None,
-        position: int = None,
-        **_,
-    ) -> None:
-        """Play the media linked to the uri provided on the device requested.
-
-        Args:
-            device_id(str): The spotify device id to play media on
-            context_uri(str, optional): The uri of the media to play.
-                Defaults to None.
-            uris(list[str], optional): List of uris to play in a
-                custom context. Defaults to None.
-            offset(int | str, optional): Where to start within the
-                context. An int is treated as a track position; a track
-                uri string starts at that track directly (avoids having
-                to resolve its position).
-            position(int, optional): The position in the song to start
-                at the media.
-
-        Raises:
-            - PlaybackError: raised when spotipy raises an error while
-                trying to start playback
-        """
-        await self.async_ensure_tokens_valid()
-
-        if context_uri is None and (uris == [] or uris is None):
-            LOGGER.info("transfering playback to device `%s`", device_id)
-            try:
-                await self.hass.async_add_executor_job(
-                    self.apis["public"].transfer_playback,
-                    device_id,
-                    True,
-                )
-                return
-            except SpotifyException as exc:
-                raise PlaybackError(exc.msg) from exc
-
-        LOGGER.info(
-            "Starting playback of `%s` on device `%s`", context_uri, device_id
-        )
-
-        if offset is not None:
-            if isinstance(offset, str):
-                offset = {"uri": offset}
-            else:
-                offset = {"position": offset}
-
-        if position is not None:
-            position = int(position * 1000)
-
-        if context_uri is not None and context_uri.startswith("spotify:track:"):
-            uris = [context_uri]
-            context_uri = None
-
-        start_args = (
-            self.apis["public"].start_playback,
-            device_id,
-            context_uri,
-            uris,
-            offset,
-            position,
-        )
-
-        try:
-            await self.hass.async_add_executor_job(*start_args)
-        except SpotifyException as exc:
-            if exc.http_status != 404:
-                raise PlaybackError(exc.msg) from exc
-
-            # A Spotify Connect device that just came online (e.g. a
-            # librespot device) may not be registered yet when we issue
-            # the play command, so Spotify answers 404 "Device not found".
-            # Wait for it to appear and retry once before giving up,
-            # instead of surfacing an error the user cannot act on.
-            LOGGER.info(
-                "Device `%s` not found on first play attempt; waiting for "
-                "it to become available",
-                device_id,
-            )
-
-            try:
-                await self.async_wait_for_device(device_id)
-                await self.hass.async_add_executor_job(*start_args)
-            except TimeoutError as timeout_exc:
-                raise PlaybackError(
-                    f"Device `{device_id}` is not available on Spotify "
-                    "Connect."
-                ) from timeout_exc
-            except SpotifyException as retry_exc:
-                raise PlaybackError(retry_exc.msg) from retry_exc
-
-    async def async_shuffle(
-        self,
-        shuffle: bool,
-        device_id: str,
-    ):
-        """Sets the shuffle mode for a device.
-
-        Args:
-            shuffle(bool): Sets the shuffle mode to True or False
-                based on the value provided
-            device_id(str): the device to set the shuffle mode on
-        """
-        await self.async_ensure_tokens_valid()
-
-        LOGGER.info(
-            "Setting shuffle to %s on device `%s`", str(shuffle), device_id
-        )
-
-        await self.hass.async_add_executor_job(
-            self.apis["public"].shuffle,
-            shuffle,
-            device_id,
-        )
-
-    async def async_liked_songs_count(self) -> int:
-        """Returns the number of linked songs for an account."""
-        await self.async_ensure_tokens_valid()
-        return await self._async_get_count(
-            self.apis["public"].current_user_saved_tracks,
-        )
-
-    async def async_liked_songs(self, force: bool = False) -> list[str]:
-        """Retrieves the list of uris of songs in the user liked songs."""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Getting saved tracks for account `%s`", self.name)
-
-        dataset = self._datasets["liked_songs"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing liked songs dataset")
-
-                liked_songs = await self._async_pager(
-                    self.apis["public"].current_user_saved_tracks,
-                )
-
-                dataset.update(liked_songs)
-            else:
-                LOGGER.debug("Using cached liked songs dataset")
-
-        return self.liked_songs
-
-    async def async_like_media(self, uris: list[str]):
-        """Adds a list of uris to the user's liked songs."""
-        await self.async_ensure_tokens_valid()
-
-        dataset = self._datasets["liked_songs"]
-
-        # Force expire the liked_songs dataset
-        async with dataset.lock:
-            dataset.expires_at = 0
-            LOGGER.debug("Expired liked_songs dataset after adding new likes")
-
-            await self.hass.async_add_executor_job(
-                self.apis["public"].current_user_saved_tracks_add,
-                uris,
-            )
-
-    async def async_repeat(
-        self,
-        state: str,
-        device_id: str,
-    ):
-        """Sets the repeat mode for a device.
-
-        Args:
-            state(str): Sets the repeat mode for the device
-            device_id(str): the device to set the repeat mode
-        """
-        await self.async_ensure_tokens_valid()
-
-        LOGGER.info(
-            "Setting repeat state to %s on device `%s`", str(state), device_id
-        )
-
-        await self.hass.async_add_executor_job(
-            self.apis["public"].repeat,
-            state,
-            device_id,
-        )
-
-    async def async_set_volume(
-        self,
-        volume: int,
-        device_id: str,
-    ):
-        """Sets the volume level for a device
-
-        Args:
-            - volume(int): The percentage of volume to set
-            - device_id(str): the device to set the repeat mode
-        """
-        await self.async_ensure_tokens_valid()
-
-        LOGGER.info("Setting volume to %d%% for device `%s`", volume, device_id)
-
-        await self.hass.async_add_executor_job(
-            self.apis["public"].volume,
-            volume,
-            device_id,
-        )
-
-    async def async_categories(
-        self,
-        force: bool = False,
-        limit: int = None,
-    ) -> list[dict]:
-        """Fetches the categories available for the account"""
-        await self.async_ensure_tokens_valid()
-        LOGGER.debug("Getting Browse Categories for account `%s`", self.name)
-
-        dataset = self._datasets["categories"]
-
-        async with dataset.lock:
-            if force or dataset.is_expired():
-                LOGGER.debug("Refreshing Browse Categories dataset")
-
-                categories = await self._async_pager(
-                    self.apis["public"].categories,
-                    prepends=[self.country, None],
-                    sub_layer="categories",
-                    max_items=limit,
-                )
-
-                dataset.update(categories)
-            else:
-                LOGGER.debug("Using cached Browse Categories dataset")
-
-        return self.categories
-
-    async def _async_get_count(
-        self,
-        function: callable,
-        prepends: list = None,
-        appends: list = None,
-        sub_layer: str = None,
-        max_items: int = None,
-    ) -> int:
-        """Returns the number of item in a specific pagination
-
-        Args:
-            - function(callable): the function to call to retrieve
-                content. Must be able to take a `limit` and `offset`
-                arguments.
-            - preppends: arguments to pass to the function on
-                each call before the limit and offset
-            - appends: arguments to pass to the function on
-                each call after the limit and offset
-            - sub_layer(str, optional): sub key in the response
-                containing the pagination. Use the response as a
-                pagination if None. Defaults to None.
-
-        Returns:
-            - int: the number of items in the pagination
-        """
-        prepends = [] if prepends is None else prepends
-        appends = [] if appends is None else appends
-        arguments = [*prepends, 1, 0, *appends]
-
-        result = await self.hass.async_add_executor_job(
-            function,
-            *arguments,
-        )
-
-        if sub_layer is not None:
-            result = result[sub_layer]
-
-        return result["total"]
-
-    async def _async_pager(
-        self,
-        function: callable,
-        prepends: list = None,
-        appends: list = None,
-        limit: int = 50,
-        sub_layer: str = None,
-        max_items: int = None,
-    ) -> list[dict]:
-        """Retrieves data from an api endpoint using a paging
-        generator logic
-
-        Args:
-            - function(callable): the function to call to retrieve
-                content. Must be able to take a `limit` and `offset`
-                arguments.
-            - preppends: arguments to pass to the function on
-                each call before the limit and offset
-            - appends: arguments to pass to the function on
-                each call after the limit and offset
-            - limit(int, optional): the maximum number of items to
-                retrieve in a single call, defaults to 50
-            - sub_layer(str, optional): sub key in the response
-                containing the pagination. Use the response as a
-                pagination if None. Defaults to None.
-            - max_items(int, optional): the maximum number of items to
-                retrieve. Retrieve all items if None. Defaults to None.
-
-        Returns:
-            - Generator[list[dict], None, None]
-        """
-        offset = 0
-        items = []
-        total = max_items
-        prepends = [] if prepends is None else prepends
-        appends = [] if appends is None else appends
-
-        while total is None or len(items) < total:
-            arguments = [*prepends, limit, offset, *appends]
-
-            result = await self.hass.async_add_executor_job(
-                function, *arguments
-            )
-
-            if sub_layer is not None:
-                result = result[sub_layer]
-
-            if total is None:
-                total = result["total"]
-
-            current_items = result["items"]
-
-            if (delta := total - (len(items) + len(result["items"]))) < 0:
-                current_items = current_items[:delta]
-
-            items.extend(current_items)
-            offset = len(items)
-
-        items = [x for x in items if x is not None]
-
-        return items
-
-    async def async_add_to_queue(self, uri: str, device_id: str = None):
-        """Adds an item to the playback queue
-
-        Args:
-            - uri(str): the spotify item to add to the queue
-            - device_id(str, optional): The id of the device this
-                command is targeting. If None, the account's currently
-                active device is the target. Defaults to None.
-        """
-        await self.async_ensure_tokens_valid()
-
-        try:
-            await self.hass.async_add_executor_job(
-                self.apis["public"].add_to_queue,
-                uri,
-                device_id,
-            )
-        except SpotifyException as exc:
-            raise PlaybackError(
-                f"Could not add `{uri}` to device `{device_id}`"
-            ) from exc
-
-    @staticmethod
-    async def async_from_config_entry(
-        hass: HomeAssistant,
-        entry: ConfigEntry,
-    ) -> "SpotifyAccount":
-        """Builds a Spotify Account from the HA config entry.
-
-        Args:
-            hass(HomeAssistant): the HomeAssistant Instance object
-            entry(ConfigEntry): the config entry for the spotify
-                account being setup
-        Returns:
-            SpotifyAccount: A spotify account from the api config in
-                the config entry
-        """
-        hass = ensure_default_data(hass, entry.entry_id)
-        domain_data = hass.data[DOMAIN]
-
-        account = domain_data[entry.entry_id].get("account")
-
-        if account is not None:
-            LOGGER.debug(
-                "Providing preexisting account for entry `%s`", entry.entry_id
-            )
-            return account
-
-        oauth_implementation = await async_get_config_entry_implementation(
-            hass=hass,
-            config_entry=entry,
-        )
-
-        public_session = PublicSession(hass, entry, oauth_implementation)
-        await public_session.async_ensure_token_valid()
-
-        private_session = DesktopSession(hass, entry)
-        await private_session.async_ensure_token_valid()
-
-        account = SpotifyAccount(
-            entry_id=entry.entry_id,
-            hass=hass,
-            public_session=public_session,
-            private_session=private_session,
-            **entry.options,
-        )
-
-        await account.async_ensure_tokens_valid(force_entry_update=True)
-
-        await account.async_profile()
-
-        LOGGER.debug(
-            "Adding entry `%s` to spotcast data entries",
-            entry.entry_id,
-        )
-        hass.data[DOMAIN][entry.entry_id]["account"] = account
-
-        return account
